move generic dataset / image collection tests
