+++ tests/unit/data/test_nprgeneral.py
@@ -1,8 +1,5 @@
-import pytest
-
+from pystiche.data import DownloadableImageCollection
 from pystiche_papers import data
-
-from tests.asserts import assert_image_downloads_correctly, assert_image_is_downloadable
 
 
 def test_NPRgeneralLicense_smoke():
@@ -10,16 +7,5 @@ def test_NPRgeneralLicense_smoke():
     assert isinstance(repr(image.license), str)
 
 
-@pytest.mark.slow
-def test_NPRgeneral_smoke(subtests):
-    for name, image in data.NPRgeneral():
-        with subtests.test(name=name):
-            assert_image_is_downloadable(image)
-
-
-@pytest.mark.large_download
-@pytest.mark.slow
-def test_NPRgeneral_images(subtests):
-    for name, image in data.NPRgeneral():
-        with subtests.test(name=name):
-            assert_image_downloads_correctly(image)
+def test_NPRgeneral_smoke():
+    assert isinstance(data.NPRgeneral(), DownloadableImageCollection)
